Editing the code for more correct work and compactness. Adding a file ItemList.jsx.
