+++ screens/Market/MarketScreen.jsx
@@ -1,23 +1,13 @@
-import React, { useContext, useEffect, useState } from "react";
-import { Text, FlatList, TextInput, TouchableOpacity, View, ScrollView, } from 'react-native';
-import Item from './Item';
-import { DNAdataContext } from '../../Data/data';
+import React from "react";
+import { Text,TextInput, TouchableOpacity, View, ScrollView, } from 'react-native';
 import styles from './style';
-import MarketSkeletonItem from "../../components/MarketSkeleton";
+import ItemList from "./ItemList";
 
-const MarketScreen = ({ navigation }) => {
-  const { marketData } = useContext(DNAdataContext);
-  const [isLoading , setIsLoading]= useState(true);
+const MarketScreen = () => {
 
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsLoading(false)
-    }, 3000);
-    return () => clearTimeout(timer);
-  }, []);
 
   return (
-    <ScrollView style={styles.marketScreen}>
+    <ScrollView style={styles.market}>
       <View style={styles.marketScreenContainer}>
         <View style={styles.headerContainer}>
           <View style={styles.headerButtonContainer}>
@@ -50,58 +40,20 @@ const MarketScreen = ({ navigation }) => {
         <Text style={styles.hotDealsText}>Hot deals</Text>
       </View>
 
-      <View style={styles.itemsContainer}>
-
-          <FlatList
-            numColumns={3}
-            data={marketData}
-            key={item => item.id}
-            keyExtractor={item => item.id}
-            renderItem={({ item }) => {
-              return  isLoading ? <MarketSkeletonItem/>
-                 : <Item item={item} navigation={navigation} />
-              ;
-            }}
-          />
-
-      </View>
+      <ItemList/>
 
       <View style={styles.hotDealsContainer}>
         <Text style={styles.hotDealsText}>Trending</Text>
       </View>
 
-      <View style={styles.itemsContainer}>
-        <FlatList
-          numColumns={3}
-          data={marketData}
-          key={item => item.id}
-          keyExtractor={item => item.id}
-          renderItem={({ item }) => {
-            return  isLoading ? <MarketSkeletonItem/>
-              : <Item item={item} navigation={navigation} />
-              ;
-          }}
-        />
-      </View>
+      <ItemList/>
 
       <View style={styles.hotDealsContainer}>
         <Text style={styles.hotDealsText}>Deals</Text>
       </View>
 
-      <View style={styles.itemsContainer}>
-        <FlatList
-          numColumns={3}
-          data={marketData}
-          key={item => item.id}
-          keyExtractor={item => item.id}
-          renderItem={({ item }) => {
-            return  isLoading ? <MarketSkeletonItem/>
-              : <Item item={item} navigation={navigation} />
-              ;
-          }}
-        />
+      <ItemList/>
 
-      </View>
     </ScrollView>
   );
 };
